price-pal-amazon-tracker-main/backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
pydantic[email]==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
APScheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
